// manual-scripts/fetch-onchain-metadata/types.ts
export type StringPublicKey = string;
export class Metadata {
	key: MetadataKey;
	updateAuthority: StringPublicKey;
	mint: StringPublicKey;
	data: Data;
	// Immutable, once flipped, all sales of this metadata are considered secondary.
	primarySaleHappened: boolean;
	// Whether or not the data struct is mutable, default is not
	isMutable: boolean;
	/// nonce for easy calculation of editions, if present
	editionNonce: number | null;
	/// Since we cannot easily change Metadata, we add the new DataV2 fields here at the end.
    token_standard: TokenStandard | null;
    /// Collection
    collection: Collection | null;
    /// Uses
    uses: Uses | null;

	// set lazy
	masterEdition?: StringPublicKey;
	edition?: StringPublicKey;

	constructor(args: {
		updateAuthority: StringPublicKey;
		mint: StringPublicKey;
		data: Data;
		primarySaleHappened: boolean;
		isMutable: boolean;
		editionNonce: number | null;
		token_standard: TokenStandard | null;
    	collection: Collection | null;
    	uses: Uses | null;
	}) {
		this.key = MetadataKey.MetadataV1;
		this.updateAuthority = args.updateAuthority;
		this.mint = args.mint;
		this.data = args.data;
		this.primarySaleHappened = args.primarySaleHappened;
		this.isMutable = args.isMutable;
		this.editionNonce = args.editionNonce;
		this.token_standard = args.token_standard;
		this.collection = args.collection;
		this.uses = args.uses;
	}
};
export class Data {
	name: string;
	symbol: string;
	uri: string;
	sellerFeeBasisPoints: number;
	creators: Creator[] | null;
	constructor(args: {
		name: string;
		symbol: string;
		uri: string;
		sellerFeeBasisPoints: number;
		creators: Creator[] | null;
	}) {
		this.name = args.name;
		this.symbol = args.symbol;
		this.uri = args.uri;
		this.sellerFeeBasisPoints = args.sellerFeeBasisPoints;
		this.creators = args.creators;
	}
};
export class Creator {
	address: StringPublicKey;
	verified: boolean;
	share: number;

	constructor(args: {
		address: StringPublicKey;
		verified: boolean;
		share: number;
	}) {
		this.address = args.address;
		this.verified = args.verified;
		this.share = args.share;
	}
};
export class Uses  { // 17 bytes + Option byte
    use_method: UseMethod; //1
    remaining: number; //8
    total: number; //8

	constructor(args:{
		use_method: UseMethod;
		remaining: number;
		total: number;
	}){
		this.use_method = args.use_method;
		this.remaining = args.remaining;
		this.total = args.total;
	}
};
export class Collection {
	verified: boolean;
    key: StringPublicKey;

	constructor(args:{
		verified: boolean;
    	key: StringPublicKey;
	}){
		this.verified = args.verified;
		this.key = args.key;
	}
};
export enum TokenStandard {
    NonFungible ,  // This is a master edition
    FungibleAsset , // A token with metadata that can also have attrributes
    Fungible ,     // A token with simple metadata
    NonFungibleEdition ,      // This is a limited edition
};
export enum UseMethod {
    Burn ,
    Multiple ,
    Single ,
};
export enum MetadataKey {
	Uninitialized = 0,
	MetadataV1 = 4,
	EditionV1 = 1,
	MasterEditionV1 = 2,
	MasterEditionV2 = 6,
	EditionMarker = 7,
};